#5 quantization: ship fp16 code_embedder (half size, lossless)
fp16 quantization (flywheel/quantize.py): 4.41MB -> 2.21MB. Gated: AUC delta fp32 vs fp16 =
0.0002 (within noise) -> adopted. code_embedder._load upcasts fp16->fp32 (numpy compute stays
fp32; the win is FILE SIZE for clone/deploy, not speed — no native fp16 BLAS in numpy). Manifest
updated (dtype float16, quantized=true, new sha). The other half of #5 (GraphCodeBERT/GNN-AST
over dataflow graphs) stays a bigger seed — multi-day build, deferred.
Suite 309 green; weights verify ok.

+++ mmorch/code_embedder.py
@@ -31,7 +31,8 @@ def _load():
         return _STATE
     d = np.load(_NPZ)
     vocab = json.loads(_VOCAB.read_text(encoding="utf-8"))
-    _STATE = {"w": {k: d[k] for k in d.files}, "vocab": vocab}
+    # upcast a float32 (los pesos pueden venir cuantizados fp16; el compute numpy es fp32)
+    _STATE = {"w": {k: d[k].astype(np.float32) for k in d.files}, "vocab": vocab}
     return _STATE
